Little reference for schema creation
I'm really green at all of this, just trying to find the most efficient way to get my own schema down in a maintainable way. Just thought maybe a more comprehensive listing of the keywords to use, and how to use them, might be useful for others trying to use this handy tool. I'm just getting into this today and will not claim to have an intuitive grasp of all the code, but this was my best understanding of what should be done. Clearly would be open to critique and corrections.